Fix Xcode build
Xcode and VisualStudio are multi-configuration CMake backends, which means that the install path can change from build to build, and CMAKE_RUNTIME_OUTPUT_DIRECTORY is just the common point between them. This changes the Mac bundle install script to not have the bundle path hard-coded, as well as special-casing Xcode and adding the configuration to the bundle path.